Modified NotificationObject
Added a new method: createNotificationPictures() that should allow for picture notifications. Also created Javadoc comments for the NotificationObject methods and constructor.
Modified the String.format() in ProgressBarActivity to follow the device's Locale settings. This way the message will be formatted properly.
Modified the MainActivity so that when the button is clicked, it will bring us to the proper ProgressBarActivity, since I have removed the extra ProgressBar_MainActivity.
Some extra .idea stuff that I'm uncertain about, but seems like it doesn't hurt the project.