Further Update to UI of ProgressBarActivity
Deleted EventAttendanceActivity since that is not what I was supposed to do.
Update the UI so that it looks closer to the User Story/Story Board.
Still need some of the functionalities to switch to other activities.
Added a new ManageMilestoneDialog, this will be a popup dialog that allows Organisers to manage their milestones.