Major Update
Added a new TextDrawable class that will draw a Letter to become the user's profile picture if they wish to delete their photo.
(Won't change their default pfp if they delete it because it'll just recreate it)
Allow the user to change their profile picture (pfp) to a picture in their device.
Need to implement Firebase functionality